Refactor: Organize scripts into a dedicated folder
- Moved `lan_identify.py` to `scripts/lan_identify.py`
- Moved `osv_upload.py` to `scripts/osv_upload.py`
- Removed unused file `VulGPT_OSV/OSV&NEO4J.py`
